chore: WIP fix babel configs
