+++ packages/polymath-investor/.babelrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  presets: ['react-app'],
+};
