--- packages/polymath-investor/babel.config.js
@@ -1,7 +1,0 @@
-const base = require('../../config/babel.config.js');
-
-module.exports = {
-  babelrcRoots: ['../packages/*'],
-  presets: ['react-app'],
-  plugins: [...base.plugins],
-};
