+++ packages/polymath-investor/package.json
@@ -9,7 +9,7 @@
     "typecheck": "flow --show-all-branches"
   },
   "devDependencies": {
-    "@babel/core": "7.0.0",
+    "@babel/core": "7.1.0",
     "@babel/node": "7.0.0",
     "@babel/plugin-proposal-class-properties": "7.0.0",
     "@babel/plugin-proposal-export-default-from": "7.0.0",
