+++ packages/polymath-issuer/.babelrc.js
@@ -1,0 +1,4 @@
+console.log('Loading issuer config');
+module.exports = {
+  presets: ['react-app'],
+};
