+++ packages/polymath-issuer/config/webpack.config.dev.js
@@ -209,6 +209,7 @@ module.exports = {
               {
                 loader: require.resolve('babel-loader'),
                 options: {
+                  configFile: '../../config/babel.config.dev.js',
                   // This is a feature of `babel-loader` for webpack (not Babel itself).
                   // It enables caching results in ./node_modules/.cache/babel-loader/
                   // directory for faster rebuilds.
@@ -235,6 +236,7 @@ module.exports = {
               {
                 loader: require.resolve('babel-loader'),
                 options: {
+                  configFile: '../../config/babel.config.dev.js',
                   babelrc: false,
                   compact: false,
                   presets: [
