+++ packages/polymath-issuer/package.json
@@ -53,7 +53,7 @@
     "whatwg-fetch": "2.0.4"
   },
   "devDependencies": {
-    "@babel/core": "7.0.0",
+    "@babel/core": "7.1.0",
     "@babel/plugin-proposal-class-properties": "7.0.0",
     "@babel/plugin-proposal-export-default-from": "7.0.0",
     "@babel/plugin-syntax-async-generators": "7.0.0",
