+++ packages/polymath-js/.babelrc.js
@@ -1,12 +1,10 @@
-const base = require('../../config/babel.config.js');
+console.log('Loading js config...');
 module.exports = {
-  babelrcRoots: ['../packages/*'],
   presets: ['@babel/flow', '@babel/env'],
   plugins: [
-    ...base.plugins,
     '@babel/plugin-proposal-object-rest-spread',
     '@babel/plugin-transform-runtime',
     '@babel/plugin-syntax-async-generators',
-    '@babel/plugin-transform-regenerator',
+    // '@babel/plugin-transform-regenerator',
   ],
 };
