+++ packages/polymath-js/package.json
@@ -27,8 +27,8 @@
   },
   "homepage": "https://github.com/PolymathNetwork/polymath-apps#readme",
   "devDependencies": {
-    "@babel/cli": "7.0.0",
-    "@babel/core": "7.0.0",
+    "@babel/cli": "7.1.0",
+    "@babel/core": "7.1.0",
     "@babel/node": "7.0.0",
     "@babel/plugin-proposal-export-default-from": "7.0.0",
     "@babel/plugin-proposal-object-rest-spread": "7.0.0",
