+++ packages/polymath-offchain/.babelrc.js
@@ -1,9 +1,6 @@
-const base = require('../../config/babel.config.js');
 module.exports = {
-  babelrcRoots: ['../packages/*'],
-  presets: ['@babel/flow', '@babel/env'],
+  presets: ['@babel/flow', '@babel/env', '@babel/react'],
   plugins: [
-    ...base.plugins,
     '@babel/plugin-proposal-object-rest-spread',
     '@babel/plugin-transform-runtime',
     '@babel/plugin-syntax-async-generators',
