+++ packages/polymath-offchain/babel.register.js
@@ -1,0 +1,5 @@
+require('@babel/register')({
+  cwd: `${__dirname}`,
+  configFile: `${__dirname}/../../config/babel.config.dev.js`,
+  ignore: [/node_modules/],
+});
