+++ packages/polymath-shared/package.json
@@ -16,8 +16,8 @@
     "test": "jest"
   },
   "devDependencies": {
-    "@babel/core": "^7.1.2",
-    "@babel/node": "7.0.0",
+    "@babel/core": "7.1.0",
+    "@babel/cli": "7.1.0",
     "@babel/plugin-proposal-class-properties": "7.0.0",
     "@babel/plugin-transform-regenerator": "^7.0.0",
     "@babel/plugin-transform-runtime": "7.0.0",
