+++ packages/polymath-shared/src/utils/index.js
@@ -7,8 +7,8 @@
  *
  * @returns {string} an URL without the trailing slash
  */
-const normalizeURL = url => url.replace(/\/$/, '');
+export const normalizeURL = (url: string) => url.replace(/\/$/, '');
 
-module.exports = {
-  normalizeURL,
-};
+// module.exports = {
+//   normalizeURL,
+// };
