+++ packages/polymath-ui/.babelrc.js
@@ -1,0 +1,5 @@
+console.log('Loading ui config...');
+module.exports = {
+  presets: ['@babel/env', '@babel/flow', '@babel/react'],
+  plugins: ['@babel/plugin-transform-runtime'],
+};
